UPDATE [18/02]
- DataSet copies and other copies removed

- started implementing interface for model

- need to re organize and clean k folds
