+++ include/PredictionModel/BaggingModel.hpp
@@ -1,12 +1,13 @@
 #ifndef BAGGING_MODEL_H_
 #define BAGGING_MODEL_H_
 
-#include "DataSet.hpp"
+//#include "DataSet.hpp"
+#include "IModel.hpp"
 #include "ICriteria.hpp"
 #include "IOperator.hpp"
 #include "RandomForest.hpp"
 
-class BaggingModel {
+class BaggingModel : IModel {
 
 private:
   // Parameters
@@ -19,27 +20,27 @@ public:
   // Constructor
 
   BaggingModel() = delete;
-  BaggingModel(std::string split_Metric, std::string split_Criteria,
-               int max_Depth);
+  BaggingModel(const std::string &split_Metric,
+               const std::string &split_Criteria, int max_Depth);
 
   // Getter
-  
+
   int get_Depth();
   int get_Trees_Number();
-  std::map<int, DecisionTree> get_Forest() const;
+  const std::map<int, DecisionTree> &get_Forest();
 
   // Setters
 
-  void set_Metric(std::string metric);
-  void set_Criteria(std::string criteria);
+  void set_Metric(const std::string &metric);
+  void set_Criteria(const std::string &criteria);
 
   // Destructor
 
-  ~BaggingModel();
+  ~BaggingModel() override;
 
   // Methods
   void aggregate_Forest(const std::map<int, DecisionTree> &forest);
-  void train(const DataSet &data);
+  void train(const DataSet &data) override;
   void train(const DataSet &data, int n);
   std::vector<double> predict(const DataSet &datas);
 };
