+++ include/Tools/Serializer.hpp
@@ -2,89 +2,35 @@
 #include <iostream>
 #include <map>
 
-#include <DecisionTree.hpp>
 #include <Answers.hpp>
+#include <DecisionTree.hpp>
 #include <boost/serialization/map.hpp>
 #include <boost/serialization/vector.hpp>
 
 namespace Serializer {
 
 //
-void serialize_Node(const TreeNode &node, const std::string &filename) {
-  std::ofstream ofs(filename, std::ios::binary);
-  boost::archive::binary_oarchive oa(ofs);
-  oa << node;
-}
+void serialize_Node(const TreeNode &node, const std::string &filename);
 
 //
-TreeNode deserialize_Node(const std::string &filename) {
-  std::ifstream ifs(filename, std::ios::binary);
-
-  if (!ifs.is_open()) {
-    std::cerr << "Error: Unable to open file '" << filename << "'" << std::endl;
-    return TreeNode();
-  }
-
-  boost::archive::binary_iarchive ia(ifs);
-
-  TreeNode node;
-  ia >> node;
-
-  return node;
-}
+TreeNode deserialize_Node(const std::string &filename);
 
 //
-void serialize_Tree(const DecisionTree &tree, const std::string &filename) {
-  std::ofstream ofs(filename, std::ios::binary);
-  boost::archive::binary_oarchive oa(ofs);
-  oa << tree;
-}
+void serialize_Tree(const DecisionTree &tree, const std::string &filename);
 
 //
-DecisionTree deserialize_Tree(const std::string &filename) {
-  std::ifstream ifs(filename, std::ios::binary);
-
-  if (!ifs.is_open()) {
-    std::cerr << "Error: Unable to open file '" << filename << "'" << std::endl;
-    return DecisionTree();
-  }
-
-  boost::archive::binary_iarchive ia(ifs);
-
-  DecisionTree tree;
-  ia >> tree;
-
-  return tree;
-}
+DecisionTree deserialize_Tree(const std::string &filename);
 
 //
-void serialize_Answers(const Answers &ans, const std::string &filename) {
-  std::ofstream ofs(filename, std::ios::binary);
-  boost::archive::binary_oarchive oa(ofs);
-  oa << ans;
-}
+void serialize_Answers(const Answers &ans, const std::string &filename);
 
 //
-Answers deserialize_Answers(const std::string &filename) {
-  std::ifstream ifs(filename, std::ios::binary);
-
-  if (!ifs.is_open()) {
-    std::cerr << "Error: Unable to open file '" << filename << "'" << std::endl;
-    return Answers();
-  }
-
-  boost::archive::binary_iarchive ia(ifs);
-
-  Answers ans;
-  ia >> ans;
-
-  return ans;
-}
+Answers deserialize_Answers(const std::string &filename);
 
 //
 template <typename Key, typename Value>
 void serialize_Map(const std::map<Key, Value> &inputMap,
-                  const std::string &filename) {
+                   const std::string &filename) {
   std::ofstream ofs(filename, std::ios::binary);
   boost::archive::binary_oarchive oa(ofs);
   oa << inputMap;
@@ -109,11 +55,10 @@ std::map<Key, Value> deserialize_Map(const std::string &filename) {
 }
 
 // Serialize function to convert arguments to a string and then serialize them
-template<typename... Args>
-std::string serialize_Data(const Args&... args) {
-    std::stringstream ss;
-    (ss << ... << args); // Serialize each argument into the stringstream
-    return ss.str();
+template <typename... Args> std::string serialize_Data(const Args &...args) {
+  std::stringstream ss;
+  (ss << ... << args); // Serialize each argument into the stringstream
+  return ss.str();
 }
 
 } // namespace Serializer
