+++ src/Criterias/Histogram.cpp
@@ -27,9 +27,10 @@ Outputs    : Object of Histogram class
 */
 Histogram::Histogram(int x) {
   if (x <= 0) {
-    errno = EINVAL;
-    perror("Can't compute split criterias with x <= 0\n x is set to 5");
-    this->number_Of_Bins = 25;
+    /* errno = EINVAL;
+    perror("Can't compute split criterias with x <= 0\n x is set to 5"); */
+    std::cerr << "Can't compute split criterias with x <= 0\n x is set to 10\n";
+    this->number_Of_Bins = 10;
   } else {
     this->number_Of_Bins = x;
   }
@@ -71,8 +72,11 @@ Outputs    : vector<double>
 */
 std::vector<double> Histogram::compute(const std::vector<double> &list) const {
 
-  auto min = std::min_element(std::execution::par_unseq, list.begin(), list.end());
-  auto max = std::max_element(std::execution::par_unseq, list.begin(), list.end());
+  /* auto min = std::min_element(std::execution::par_unseq, list.begin(), list.end());
+  auto max = std::max_element(std::execution::par_unseq, list.begin(), list.end()); */
+
+  auto min = std::min_element(list.begin(), list.end());
+  auto max = std::max_element(list.begin(), list.end());
 
   std::vector<double> res(this->number_Of_Bins, 0.0);
